src/main/java/com/loofah/graph/api/services/CategoryService.java
package com.loofah.graph.api.services;

import com.loofah.graph.api.models.database.Category;
import com.loofah.graph.api.repositories.CategoryRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import java.util.List;

@Component
public class CategoryService {

    private CategoryRepository categoryRepository;

    @Autowired
    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    public Category getById(String id) {
        return categoryRepository.findById(id).get();
    }

    public List<Category> getAll() {
        return categoryRepository.findAll();
    }
}
